applied DistributedFixture to first batch of tests as a trial
